Move 87FB20 and cleanup
