+++ src/camera_check/87A320.c
@@ -1,3 +1,4 @@
 #include "common.h"
+#include "camera_check.h"
 
 #pragma GLOBAL_ASM("asm/nonmatchings/camera_check/87A320/func_camera_check_801DCB30.s")
