+++ src/camera_check/87A330.c
@@ -1,60 +1,5 @@
 #include "common.h"
-
-typedef struct SiennaHaddock {
-    char pad[0x48];
-    SObj* var_48;
-} SiennaHaddock;
-
-typedef struct PurpleMackerel {
-    /* 0x00 */ SiennaHaddock* var_0;
-    /* 0x04 */ char pad_04[0x1C];
-} PurpleMackerel; // size: 0x20
-
-typedef struct TealFlounder {
-    GObj* unk0;
-    SObj* unk4;
-    SObj* unk8;
-    SObj* unkC;
-    SObj* unk10;
-    SObj* unk14;
-    SObj* unk18;
-    SObj* unk1C;
-} TealFlounder;
-
-extern Sprite D_camera_check_801FCC58;
-extern Sprite D_camera_check_801FD088;
-extern Sprite D_camera_check_801FD9E0;
-extern Sprite D_camera_check_801FDDB8;
-extern Sprite D_camera_check_801FE160;
-extern Sprite D_camera_check_801FF128;
-extern Sprite D_camera_check_801FF988;
-extern Sprite D_camera_check_802004E8;
-extern Sprite D_camera_check_80208900;
-extern char D_camera_check_80208A34[];
-extern char D_camera_check_80208A3C[];
-extern TealFlounder D_camera_check_802498F0;
-extern s32 D_camera_check_802498FC;
-extern s32 D_camera_check_80249910;
-extern s32 D_camera_check_80249914;
-extern s32 D_camera_check_80249918;
-extern s32 D_camera_check_8024991C;
-extern SObj* D_camera_check_802499A8;
-extern PurpleMackerel D_camera_check_802499C0[];
-
-int func_800BF8BC_5C75C(int);
-void func_camera_check_801DCB58(int);
-void func_camera_check_801DCC0C(void);
-void func_camera_check_801DCD00(void);
-void func_camera_check_801DCF9C(int);
-void func_camera_check_801DD238(void*, s32, int);
-void func_camera_check_801DDF28(s32 arg0);
-void func_camera_check_801DE288(int);
-void func_camera_check_801DF2D8(UNK_PTR);
-int func_camera_check_801E2484(int);
-int func_camera_check_801E24D8(int);
-void func_803713D4_844B84(int);
-void func_803713EC_844B9C(int);
-void func_80374714_847EC4(int, int);
+#include "camera_check.h"
 
 s32 func_camera_check_801DCB40(void) {
     return D_camera_check_80249914;
